src/main/java/de/eb17/episodetracker/service/MemberRegistration.java
package de.eb17.episodetracker.service;

import de.eb17.episodetracker.model.Member;

import javax.ejb.Stateless;
import javax.enterprise.event.Event;
import javax.inject.Inject;
import javax.persistence.EntityManager;
import java.util.logging.Logger;

/**
 * Class for registering a member.
 *
 */
@Stateless
public class MemberRegistration {

    @Inject
    private Logger log;

    @Inject
    private EntityManager em;

    @Inject
    private Event<Member> memberEventSrc;

    public void register(Member member) throws Exception {
        log.info("Registering " + member.getSurname());
        em.persist(member);
        memberEventSrc.fire(member);
    }
}
